Publishing clickable graph to pointdata.py
