+++ cgi-bin/devtests/pyroplot_local.py
