+++ cgi-bin/live/pointdata.py
@@ -1,5 +1,11 @@
 #!/usr/bin/python
 
+import matplotlib as mpl
+# It is vital to set the matplotlib interface to one without a display
+# BEFORE loading pyplot!  This makes things work behind Apache.
+mpl.use('Agg')
+import matplotlib.pyplot as plt
+
 import pmcgi
 import os
 import pyromat as pm
@@ -7,14 +13,13 @@ import numpy as np
 import pyromat.solve as pmsolve
 import liveutils as lu
 
-
 source = '/var/www/html/live/pointdata.html'
 
 # Print the header so the page will display even if something goes wrong
 print("Content-type: text/html")
 print("")
 
-#Create the HTML page object
+# Create the HTML page object
 P = pmcgi.PMPage(source)
 
 #find critical lines in the
@@ -24,6 +29,8 @@ unitline = P.find_line('<!-- units -->')+4 #line where we begin the unit selecto
 resline = P.find_line('<!-- results -->')+1 #line where we begin the results
 errline = P.find_line('<!-- errors -->') +1 #line where we insert errors
 chartline = P.find_line('<!-- charts -->')+1 #line where we begin inserting a chart
+coordsline = P.find_line('<!-- coordinates -->')+42 #line where we begin inserting chart coords
+#coordsline = 112
 
 #Try to parse the user's inputs
 try:
@@ -70,15 +77,15 @@ sval = Tval = pval = hval = vval = xval = ''
 
 #check for cases where too few or too many properties are specified (zero specified results in defaults)
 ins = np.array([p1,T1,s1,v1,h1,x1]) #test array for counting how many properties were specified
-if (sum(ins>=0)==1) or (sum(ins>=0)>2):
+if (sum(ins>=-9000)==1) or (sum(ins>=-9000)>2):
     lu.perror(P,'Must specify exactly two properties to define a state.', errline)
-elif (sum(ins>=0)==0):
+elif (sum(ins>=-9000)==0):
     p1 = 101.325
     T1 = 300
 
 #Determine which two properties were chosen, calculate the state
 try:
-    if p1 >= 0 and s1 >= 0: #P&s
+    if p1 >= -9000 and s1 >= -9000: #P&s
         #Set the input box labels
         pval = p1
         sval = s1
@@ -91,7 +98,7 @@ try:
             d1 = F.d(T=T1, p=p1)
             h1 = F.h(T=T1, p=p1)
         v1 = 1 / d1
-    elif T1 >= 0 and s1 >= 0: #T&s
+    elif T1 >= -9000 and s1 >= -9000: #T&s
         # Set the input box labels
         Tval = T1
         sval = s1
@@ -106,7 +113,7 @@ try:
             h1, s1, d1 = F.hsd(p=p1, T=T1)
         v1 = 1 / d1
         p1 = F.p(T=T1, d=d1)
-    elif T1 >= 0 and p1 >= 0: #T&p
+    elif T1 >= -9000 and p1 >= -9000: #T&p
         #set the input box labels
         Tval = T1
         pval = p1
@@ -115,7 +122,7 @@ try:
         v1 = 1 / d1
         p1 = F.p(T=T1, d=d1)
         x1 = -1 #P&T will always fall outside the dome
-    elif T1 >= 0 and h1 >= 0: #T&h
+    elif T1 >= -9000 and h1 >= -9000: #T&h
         #set the input box labels
         Tval = T1
         hval = h1
@@ -130,7 +137,7 @@ try:
             h1, s1, d1 = F.hsd(p=p1, T=T1)
         v1 = 1 / d1
         p1 = F.p(T=T1, d=d1)
-    elif p1 >= 0 and h1 >= 0: #P&h
+    elif p1 >= -9000 and h1 >= -9000: #P&h
         #set the input box labels
         hval = h1
         pval = p1
@@ -142,7 +149,7 @@ try:
             h1, s1, d1 = F.hsd(p=p1, T=T1)
         v1 = 1 / d1
         p1 = F.p(T=T1, d=d1)
-    elif p1 >= 0 and v1 >= 0: #P&v
+    elif p1 >= -9000 and v1 >= -9000: #P&v
         #set the input box labels
         pval = p1
         vval = v1
@@ -156,7 +163,7 @@ try:
             h1, s1, d1 = F.hsd(p=p1, T=T1)
         v1 = 1 / d1
         p1 = F.p(T=T1, d=d1)
-    elif T1 >= 0 and v1 >= 0: #T&v
+    elif T1 >= -9000 and v1 >= -9000: #T&v
         #Set the input box labels
         Tval = T1
         vval = v1
@@ -176,7 +183,7 @@ try:
             h1, s1, d1, x1 = F.hsd(p=p1, T=T1, quality=True)
         v1 = 1 / d1
         p1 = F.p(T=T1, d=d1)
-    elif h1 >= 0 and v1 >= 0: #h&v
+    elif h1 >= -9000 and v1 >= -9000: #h&v
         #set the input box labels
         hval = h1
         vval = v1
@@ -192,7 +199,7 @@ try:
             h1, s1, d1 = F.hsd(p=p1, T=T1)
         v1 = 1 / d1
         p1 = F.p(T=T1, d=d1)
-    elif s1 >= 0 and v1 >= 0: #s&v
+    elif s1 >= -9000 and v1 >= -9000: #s&v
         #set the input box labels
         sval = s1
         vval = v1
@@ -208,7 +215,7 @@ try:
             h1, s1, d1 = F.hsd(p=p1, T=T1)
         v1 = 1 / d1
         p1 = F.p(T=T1, d=d1)
-    elif x1 >=0 and T1 >= 0: #T&x
+    elif x1 >=-9000 and T1 >= -9000: #T&x
         #set the input box labels
         Tval = T1
         xval = x1
@@ -216,7 +223,7 @@ try:
         p1 = F.ps(T=T1)
         h1, s1, d1 = F.hsd(T=T1, x=x1)
         v1 = 1/d1
-    elif x1 >=0 and p1 >= 0: #P&x
+    elif x1 >=-9000 and p1 >= -9000: #P&x
         #set the input box labels
         pval = p1
         xval = x1
@@ -271,11 +278,58 @@ P.insert('<h3>State Properties</h3><center>' +
          + '</center>', (resline, 0), wait=True)
 
 # Insert the cgi call to build the image
+# """<canvas id="myCanvas"  style="cursor:crosshair;background: url('/cgi-bin/live/pointdata_plot.py?id={:s}&p1={:f}&s1={:f}&up={:s}&uT={:s}&uE={:s}&uM={:s}&uV={:s}')" width="640" height="480"></canvas>""".format(
+#        species, float(p1), float(s1), up, uT, uE, uM, uV)
 P.insert(
-    '<img class="figure" src="/cgi-bin/live/pointdata_plot.py?id={:s}&p1={:f}&s1={:f}&up={:s}&uT={:s}&uE={:s}&uM={:s}&uV={:s}">'.format(
+    """/cgi-bin/live/pointdata_plot.py?id={:s}&p1={:f}&s1={:f}&up={:s}&uT={:s}&uE={:s}&uM={:s}&uV={:s}""".format(
         species, float(p1), float(s1), up, uT, uE, uM, uV),
-    (chartline, 0))
+    (chartline, 63))
 
+#insert the chartcoordinates
+
+
+def getChartCoords():
+    f = plt.figure()
+    ax = f.add_subplot(111)
+    Tt = F.triple()[0]
+    Tc = F.critical()[0]
+    temp = (Tc - Tt) * .00001
+    T = np.linspace(Tt + temp, Tc - temp, 5)
+    sL, sV = F.ss(T)
+    ax.plot(sL, T, lw=2, color='k')
+    ax.plot(sV, T, lw=2, color='k')
+
+    # Add an isobar
+    Tmin = F.Tlim()[0]
+    Tmax = F.Tlim()[1]
+    smin = F.s(T=Tmin, p=p1)
+    smax = F.s(T=Tmax, p=p1)
+    s = np.linspace(smin, smax, 5)
+    try:
+        with lu.nostdout():
+            T = F.T_s(p=p1, s=s)
+            ax.plot(s, T, lw=2, color='b', linestyle=':')
+    except:
+        pass
+
+    origin = ax.get_position()
+    originx_pct = str(origin.get_points()[0][0])
+    originy_pct = str(origin.get_points()[0][1])
+    maxx_pct = str(origin.get_points()[1][0])
+    maxy_pct = str(origin.get_points()[1][1])
+    xlim_low = str(ax.get_xlim()[0])
+    xlim_hi = str(ax.get_xlim()[1])
+    ylim_low = str(ax.get_ylim()[0])
+    ylim_hi = str(ax.get_ylim()[1])
+    P.insert(originx_pct, (coordsline, 23), wait=True)
+    P.insert(originy_pct, (coordsline+1, 23), wait=True)
+    P.insert(maxx_pct, (coordsline+2, 23), wait=True)
+    P.insert(maxy_pct, (coordsline+3, 23), wait=True)
+    P.insert(xlim_low, (coordsline+4, 23), wait=True)
+    P.insert(xlim_hi, (coordsline+5, 23), wait=True)
+    P.insert(ylim_low, (coordsline+6, 23), wait=True)
+    P.insert(ylim_hi, (coordsline+7, 23), wait=True)
+getChartCoords()
 
 #Perform the write operation
 P.insert_exec()
