--- html/live/indextest.html
@@ -1,33 +1,0 @@
-<!DOCTYPE HTML>
-
-<html>
-<head>
-    <title>Live | PYroMat</title>
-    <link rel="stylesheet" type="text/css" href="/pyromat.css">
-    <meta name="description" content="getting python">
-    <meta name="keywords" content="python,thermo,thermodynamic,property,properties">
-    <meta name="author" content="C.R. Martin">
-</head>
-<body>
-
-<iframe name="menu" id="menu" src="/menu.html"></iframe>
-
-<div id="content">
-<div class="title">
-<h1>PYroMat Live Modules</h1>
-</div>
-<div class="section">
-    <p>
-    Currently availabel PYroMat Live modules:
-    </p>
-
-    <div class="toc0"><a href="/cgi-bin/live/rankine.py">The Rankine Cycle</a></div>
-	<div class="toc0"><a href="/cgi-bin/live/pointdata.py">Single Point Property Data</a></div>
-    <div class="toc0"><a href="/cgi-bin/live/satdata.py">Saturation Property Data</a></div>
-</div>
-<div style="height:100px"></div>
-
-<iframe src="/author.html" id="contact"></iframe>
-
-</body>
-</html>
