+++ html/live/pointdata_old.html
@@ -1,0 +1,81 @@
+<!DOCTYPE HTML>
+
+<html>
+<head>
+    <title>Thermodynamic Property Data | PYroMat</title>
+    <link rel="stylesheet" type="text/css" href="/pyromat.css">
+    <meta name="description" content="getting python">
+    <meta name="keywords" content="python,thermo,thermodynamic,property,properties">
+    <meta name="author" content="J. Ranalli">
+    
+<script>
+function toggle_vis(target_id){
+    var target_obj = document.getElementById(target_id);
+    if (target_obj.style.visibility=='visible'){
+        target_obj.style.visibility = 'hidden';
+    } else {
+        target_obj.style.visibility = 'visible';
+    }
+}
+</script>
+</head>
+<body>
+
+<iframe name="menu" id="menu" src="/menu.html"></iframe>
+
+<div id="content">
+<div class="title">
+<h1>Thermodynamic Property Calculator</h1>
+</div>
+
+
+<div class="section">
+    <h2>The State Principle</h2>
+    <p>The state principle guarantees that any two independent, intensive properties are enough to fix the state.
+    Fixing the state allows any additional intensive property to be found. Try entering any pair of properties below.
+</div>
+
+<form action="../../cgi-bin/live/pointdata_old.py" method="POST">
+    <fieldset>
+        <legend><b>State Parameters</b></legend>
+<!-- inputs -->
+        <div class="lcol">
+        <table id="values">
+            <tr><td>Fluid</td><td><select name="id"></select></td></tr>
+            <tr><td>Pressure</td><td><input name="p1" type="text" value=""></td></tr>
+            <tr><td>Temperature</td><td><input name="T1" type="text" value=""></td></tr>
+            <tr><td>Specific Enthalpy</td><td><input name="h1" type="text" value=""></td></tr>
+            <tr><td>Specific Entropy</td><td><input name="s1" type="text" value=""></td></tr>
+            <tr><td>Specific Volume</td><td><input name="v1" type="text" value=""></td></tr>
+            <tr><td>Quality</td><td><input name="x1" type="text" value=""></td></tr>
+
+        </table>
+        <input type="submit" value="Update">
+        </div>
+        
+<!-- units -->
+        <div class="rcol">
+        <a href="#" onClick="javascript:toggle_vis('units')">Units (Show/Hide)</a><br>
+        <table id="units" style="visibility:hidden;">
+            <tr><td>Pressure</td><td><select name="up"></select></td></tr>
+            <tr><td>Temperature</td><td><select name="uT"></select></td></tr>
+            <tr><td>Energy</td><td><select name="uE"></select></td></tr>
+            <tr><td>Matter</td><td><select name="uM"></select></td></tr>
+            <tr><td>Volume</td><td><select name="uV"></select></td></tr>
+
+
+        </table>
+        </div>
+        
+    </fieldset>
+</form>
+<!-- errors -->
+<!-- results -->
+<!-- charts -->
+<div class="section">
+
+
+</div>
+
+</body>
+</html>
